Support JWT authentication as well as basic auth
In order to support websites that use alternative authentication as opposed to
using a username/password in the URL, I've added the option to pass in a
token to be used as the value of the 'Authorization' header.

While looking at the code I noticed we did not namespace one of the
classes (`Page`), and added a spec for the `Crawl::Engine`. We are also
introducing webmock to be able to fake the whole process and make sure
we are hitting the website with the correct headers.